pyproject.toml
[tool.poetry]
name = "nonebot_plugin_setu"
version = "1.0.14"
description = "基于lolicon api的涩图插件"
authors = ["ayanamiblhx <1196818079@qq.com>"]
license = "Apache License 2.0"
homepage = "https://github.com/ayanamiblhx/nonebot_plugin_setu"
documentation = "https://github.com/ayanamiblhx/nonebot_plugin_setu/blob/main/README.md"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.21.3"
httpx-socks = "^0.7.3"
nonebot2 = "^2.0.0b2"
nonebot-adapter-onebot = "^2.0.0b1"
tqdm = "^4.61.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
